# api/openapi.yaml
openapi: 3.1.0
info:
  title: ergonomos API
  version: 0.0.0
  description: |
    REST contract for the ergonomos backend and the source of truth for the
    typed Dart client (see docs/adr/0001).

    Conventions:
      * All resource identifiers are opaque, globally-unique strings (never
        sequential integers) — see docs/adr/0004 (federation-readiness).
      * Shareable resources expose a canonical URI and a home-instance marker.
      * Errors use RFC 9457 problem details.
  license:
    name: AGPL-3.0-or-later WITH ergonomos-app-store-exception
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080
    description: Local development
paths:
  /healthz:
    get:
      summary: Liveness probe
      operationId: getHealthz
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
  /auth/register:
    post:
      summary: Register a new account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '201':
          description: Account created; an authentication token is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/Problem'
        '409':
          description: An account already exists for this email.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /me:
    get:
      summary: The currently authenticated actor
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The actor the bearer token was issued for.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Actor'
        '401':
          description: Missing or invalid authentication token.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /auth/sessions:
    post:
      summary: Sign in with email and password
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          description: Credentials accepted; an authentication token is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          description: Authentication failed (unknown email or wrong password).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /tasks:
    post:
      summary: Create a task owned by the authenticated actor
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '201':
          description: The created task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          description: Missing or invalid authentication token.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    get:
      summary: List the tasks owned by the authenticated actor
      operationId: listTasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The caller's tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Missing or invalid authentication token.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /tasks/{id}:
    get:
      summary: Retrieve one of the authenticated actor's tasks
      operationId: getTask
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Opaque task identifier.
      responses:
        '200':
          description: The requested task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Missing or invalid authentication token.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: |
            No such task is visible to the caller. Unknown id and a task owned
            by someone else both return 404 so the API cannot be used to probe
            for others' tasks.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An opaque token issued by /auth/register or /auth/sessions.
  responses:
    Problem:
      description: Request could not be processed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok
    Credentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: ada@example.org
        password:
          type: string
          format: password
          minLength: 8
    Actor:
      description: A principal that can hold relationships to objects (ADR-0004).
      type: object
      required: [id, handle, instance, local]
      properties:
        id:
          type: string
          description: Opaque, globally-unique identifier (ULID).
          example: 01J9Z3K7QF8XN2M4R5T6V7W8Y9
        handle:
          type: string
          description: Local part on this instance; "ada" for ada@example.org.
          example: ada
        instance:
          type: string
          description: Home-instance domain.
          example: localhost
        local:
          type: boolean
          example: true
    Session:
      type: object
      required: [token, actor]
      properties:
        token:
          type: string
          description: Opaque bearer token for the authenticated actor.
        actor:
          $ref: '#/components/schemas/Actor'
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          example: Write acceptance tests
    Task:
      description: A unit of work owned by an actor (ADR-0006, first-class aggregate).
      type: object
      required: [id, title, owner, createdAt]
      properties:
        id:
          type: string
          description: Opaque, globally-unique identifier (ULID).
          example: 01J9Z3K7QF8XN2M4R5T6V7W8Y9
        title:
          type: string
          example: Write acceptance tests
        owner:
          type: string
          description: Opaque id of the owning actor.
          example: 01J9Z3K7QF8XN2M4R5T6V7W8Y9
        createdAt:
          type: string
          format: date-time
          description: Creation time (UTC).
    Problem:
      description: RFC 9457 problem detail
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string